Add tags for images
Pexels as well as other sources of images often already include a list
of tags for those images. Those tags might be useful for running
additional analytics to map out the food space (content filtering as
opposed to collaborative filtering). The tag data doesn't necessarily
belong here alongside the images, but it seems to be a better place than
nowhere in case anybody wants to use it.