Create indexed decision tree for crystal bowl
Crystal bowl is our playful twist on a magical crystal ball with which
you can see into the future -- your own future with regards to what food
you're about to decide to eat that is... May your food come to you in a
crystal bowl...

Anyway, the tree-map.json file is an indexed node tree. Each element in
the array represents a node in the tree with child branches that contain
pointers (indexes) to other elements in the tree map array. For each
entry, the "names" array is an array of four food images -- the first,
left middle, right middle and last food items in that branch. The middle
terms are useful for bisecting the tree branch where left and right
middle simply mean the last/first elements of the left and right child
branches. So for instance, if we were talking about the range of
integers from 1 to 100, the first integer would be 1, the last would be
100, the left middle would be 50 and the right middle would be 51, so
the array would contain the values [ 1, 50, 51, 100 ] in that order.