Remove outdated images and update metadata
A long time ago we removed a number of images from the "menu" (the list
of images that would actively display on feed me jefferson), but we
never removed the images themselves, or their metadata so that we'd be
able to include them in any historical analysis. As a result, a lot of
those images showed up again in our attempts to manually curate tags
creating extra work for us and a bit of frustration at the bad quality
of a lot of images (which never should have shown up in the work). To
avoid having this happen again, I'm fully wiping them from our repo
(they'll still be there in history), but they shouldn't be floating
around in the directory list anymore.